Extend fuzzing tool
